requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.7
distro==1.9.0
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
filelock==3.25.2
fsspec==2026.3.0
google-auth==2.49.1
google-genai==1.70.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
kombu==5.6.2
llvmlite==0.47.0
MarkupSafe==3.0.3
more-itertools==11.0.1
mpmath==1.3.0
networkx==3.6.1
numba==0.65.0
numpy==2.4.4
openai-whisper==20250625
packaging==26.0
prompt_toolkit==3.0.52
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
python-dateutil==2.9.0.post0
regex==2026.4.4
requests==2.33.1
setuptools==81.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
sympy==1.14.0
tenacity==9.1.4
tiktoken==0.12.0
torch==2.11.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
websockets==16.0
yt-dlp==2026.3.17
